+++ src/main/java/com/imt/Api_monstres/controller/dto/input/SkillHttpDto.java
@@ -10,11 +10,11 @@ import lombok.Getter;
 @AllArgsConstructor
 public class SkillHttpDto {
     private final Integer number;
-    private final Double damage;
+    private final Integer damage;
     private final Ratio ratio;
-    private final Double cooldown;
-    private final Double lvl;
-    private final Double lvlMax;
+    private final Integer cooldown;
+    private final Integer lvl;
+    private final Integer lvlMax;
     private final Rank rank;
 }
